testing with api of exams
